+++ Changelog.md
@@ -8,6 +8,7 @@
 - Dropped PHP 8.2 Support
 - Dropped Laravel 10 Support
 - Added Laravel 12 Support
+- Added the `gtin` field to the product and master product variant forms
 
 ## 4.2.0
 ##### 2024-12-15